simple_lotto/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
tenacity==8.5.0
pydantic==2.8.2
